chore: code cleanup, remove old vue files, restructure code folders
