+++ .electron-vue/webpack.main.config.js
@@ -9,69 +9,69 @@ const webpack = require('webpack')
 const BabelMinifyPlugin = require('babel-minify-webpack-plugin');
 
 let mainConfig = {
-  entry: {
-    main: path.join(__dirname, '../src/main/index.js')
-  },
-  externals: [
-    ...Object.keys(dependencies || {})
-  ],
-  module: {
-    rules: [
-      {
-        test: /\.(js)$/,
-        enforce: 'pre',
-        exclude: /node_modules/,
-        use: {
-          loader: 'eslint-loader',
-        }
-      },
-      {
-        test: /\.js$/,
-        use: 'babel-loader',
-        exclude: /node_modules/
-      },
-      {
-        test: /\.node$/,
-        use: 'node-loader'
-      }
-    ]
-  },
-  node: {
-    __dirname: process.env.NODE_ENV !== 'production',
-    __filename: process.env.NODE_ENV !== 'production'
-  },
-  output: {
-    filename: '[name].js',
-    libraryTarget: 'commonjs2',
-    path: path.join(__dirname, '../dist/electron')
-  },
-  plugins: [
-    new webpack.NoEmitOnErrorsPlugin()
-  ],
-  resolve: {
-    extensions: ['.js', '.json', '.node']
-  },
-  target: 'electron-main'
+	entry: {
+		main: path.join(__dirname, '../src/main/index.js')
+	},
+	externals: [
+		...Object.keys(dependencies || {})
+	],
+	module: {
+		rules: [
+			{
+				test: /\.(js)$/,
+				enforce: 'pre',
+				exclude: /node_modules/,
+				use: {
+					loader: 'eslint-loader',
+				}
+			},
+			{
+				test: /\.js$/,
+				use: 'babel-loader',
+				exclude: /node_modules/
+			},
+			{
+				test: /\.node$/,
+				use: 'node-loader'
+			}
+		]
+	},
+	node: {
+		__dirname: process.env.NODE_ENV !== 'production',
+		__filename: process.env.NODE_ENV !== 'production'
+	},
+	output: {
+		filename: '[name].js',
+		libraryTarget: 'commonjs2',
+		path: path.join(__dirname, '../dist/electron')
+	},
+	plugins: [
+		new webpack.NoEmitOnErrorsPlugin()
+	],
+	resolve: {
+		extensions: ['.js', '.json', '.node']
+	},
+	target: 'electron-main'
 }
 
 /**
  * Adjust mainConfig for development settings
  */
 if (process.env.NODE_ENV !== 'production') {
-  mainConfig.plugins.push(
-    new webpack.DefinePlugin({
-      '__static': `"${path.join(__dirname, '../static').replace(/\\/g, '\\\\')}"`
-    })
-  )
+	mainConfig.plugins.push(
+		new webpack.DefinePlugin({
+			'__static': `"${path.join(__dirname, '../static').replace(/\\/g, '\\\\')}"`
+		})
+	)
 }
 
 /**
  * Adjust mainConfig for production settings
  */
 if (process.env.NODE_ENV === 'production') {
-  mainConfig.plugins.push(
-    new BabelMinifyPlugin()
-  )
+	mainConfig.plugins.push(
+		new BabelMinifyPlugin()
+	)
 }
 
 module.exports = mainConfig
