+++ .electron-vue/webpack.renderer.config.js
@@ -21,171 +21,171 @@ const HtmlWebpackPlugin = require('html-webpack-plugin')
 let whiteListedModules = ['vue']
 
 let rendererConfig = {
-  devtool: '#cheap-module-eval-source-map',
-  entry: {
-    renderer: path.join(__dirname, '../src/renderer/main.js')
-  },
-  externals: [
-    ...Object.keys(dependencies || {}).filter(d => !whiteListedModules.includes(d))
-  ],
-  module: {
-    rules: [
-      {
-        test: /\.(js|vue)$/,
-        enforce: 'pre',
-          loader: 'eslint-loader',
-        include: [
-          path.resolve(__dirname, '../src/renderer'),
-        ]
-      },
-      {
-        test: /\.css$/,
-        use: ExtractTextPlugin.extract({
-          fallback: 'style-loader',
-          use: 'css-loader'
-        })
-      },
-      {
-        test: /\.html$/,
-        use: 'vue-html-loader'
-      },
-      {
-        test: /\.js$/,
-        use: 'babel-loader',
-        exclude: /node_modules/
-      },
-      {
-        test: /\.node$/,
-        use: 'node-loader'
-      },
-      {
-        test: /\.vue$/,
-        use: {
-          loader: 'vue-loader',
-          options: {
-            extractCSS: process.env.NODE_ENV === 'production',
-            loaders: {
-              sass: 'vue-style-loader!css-loader!sass-loader?indentedSyntax=1',
-              scss: 'vue-style-loader!css-loader!sass-loader'
-            }
-          }
-        }
-      },
-      {
-        test: /\.(png|jpe?g|gif|svg)(\?.*)?$/,
-        use: {
-          loader: 'url-loader',
-          query: {
-            limit: 10000,
-            name: 'imgs/[name]--[folder].[ext]'
-          }
-        }
-      },
-      {
-        test: /\.(mp4|webm|ogg|mp3|wav|flac|aac)(\?.*)?$/,
-        loader: 'url-loader',
-        options: {
-          limit: 10000,
-          name: 'media/[name]--[folder].[ext]'
-        }
-      },
-      {
-        test: /\.(woff2?|eot|ttf|otf)(\?.*)?$/,
-        use: {
-          loader: 'url-loader',
-          query: {
-            limit: 10000,
-            name: 'fonts/[name]--[folder].[ext]'
-          }
-        }
-      }
-    ]
-  },
-  node: {
-    __dirname: process.env.NODE_ENV !== 'production',
-    __filename: process.env.NODE_ENV !== 'production'
-  },
-  plugins: [
-    new ExtractTextPlugin('styles.css'),
-    new HtmlWebpackPlugin({
-      filename: 'assistant.html',
-      template: path.resolve(__dirname, '../src/assistant.ejs'),
-      minify: {
-        collapseWhitespace: true,
-        removeAttributeQuotes: true,
-        removeComments: true
-      },
-      nodeModules: process.env.NODE_ENV !== 'production'
-        ? path.resolve(__dirname, '../node_modules')
-        : false
+	devtool: '#cheap-module-eval-source-map',
+	entry: {
+		renderer: path.join(__dirname, '../src/renderer/index.js')
+	},
+	externals: [
+		...Object.keys(dependencies || {}).filter(d => !whiteListedModules.includes(d))
+	],
+	module: {
+		rules: [
+			{
+				test: /\.(js|vue)$/,
+				enforce: 'pre',
+				loader: 'eslint-loader',
+				include: [
+					path.resolve(__dirname, '../src/renderer'),
+				]
+			},
+			{
+				test: /\.css$/,
+				use: ExtractTextPlugin.extract({
+					fallback: 'style-loader',
+					use: 'css-loader'
+				})
+			},
+			{
+				test: /\.html$/,
+				use: 'vue-html-loader'
+			},
+			{
+				test: /\.js$/,
+				use: 'babel-loader',
+				exclude: /node_modules/
+			},
+			{
+				test: /\.node$/,
+				use: 'node-loader'
+			},
+			{
+				test: /\.vue$/,
+				use: {
+					loader: 'vue-loader',
+					options: {
+						extractCSS: process.env.NODE_ENV === 'production',
+						loaders: {
+							sass: 'vue-style-loader!css-loader!sass-loader?indentedSyntax=1',
+							scss: 'vue-style-loader!css-loader!sass-loader'
+						}
+					}
+				}
+			},
+			{
+				test: /\.(png|jpe?g|gif|svg)(\?.*)?$/,
+				use: {
+					loader: 'url-loader',
+					query: {
+						limit: 10000,
+						name: 'imgs/[name]--[folder].[ext]'
+					}
+				}
+			},
+			{
+				test: /\.(mp4|webm|ogg|mp3|wav|flac|aac)(\?.*)?$/,
+				loader: 'url-loader',
+				options: {
+					limit: 10000,
+					name: 'media/[name]--[folder].[ext]'
+				}
+			},
+			{
+				test: /\.(woff2?|eot|ttf|otf)(\?.*)?$/,
+				use: {
+					loader: 'url-loader',
+					query: {
+						limit: 10000,
+						name: 'fonts/[name]--[folder].[ext]'
+					}
+				}
+			}
+		]
+	},
+	node: {
+		__dirname: process.env.NODE_ENV !== 'production',
+		__filename: process.env.NODE_ENV !== 'production'
+	},
+	plugins: [
+		new ExtractTextPlugin('styles.css'),
+		new HtmlWebpackPlugin({
+			filename: 'assistant.html',
+			template: path.resolve(__dirname, '../src/assistant.ejs'),
+			minify: {
+				collapseWhitespace: true,
+				removeAttributeQuotes: true,
+				removeComments: true
+			},
+			nodeModules: process.env.NODE_ENV !== 'production'
+				? path.resolve(__dirname, '../node_modules')
+				: false
 		}),
 		new HtmlWebpackPlugin({
-      filename: 'response.html',
-      template: path.resolve(__dirname, '../src/response.ejs'),
-      minify: {
-        collapseWhitespace: true,
-        removeAttributeQuotes: true,
-        removeComments: true
-      },
-      nodeModules: process.env.NODE_ENV !== 'production'
-        ? path.resolve(__dirname, '../node_modules')
-        : false
-    }),
-    new webpack.HotModuleReplacementPlugin(),
-    new webpack.NoEmitOnErrorsPlugin(),
-  ],
-  output: {
-    filename: '[name].js',
-    libraryTarget: 'commonjs2',
-    path: path.join(__dirname, '../dist/electron')
-  },
-  resolve: {
-    alias: {
-      '@': path.join(__dirname, '../src/renderer'),
-      'vue$': 'vue/dist/vue.esm.js'
-    },
+			filename: 'response.html',
+			template: path.resolve(__dirname, '../src/response.ejs'),
+			minify: {
+				collapseWhitespace: true,
+				removeAttributeQuotes: true,
+				removeComments: true
+			},
+			nodeModules: process.env.NODE_ENV !== 'production'
+				? path.resolve(__dirname, '../node_modules')
+				: false
+		}),
+		new webpack.HotModuleReplacementPlugin(),
+		new webpack.NoEmitOnErrorsPlugin(),
+	],
+	output: {
+		filename: '[name].js',
+		libraryTarget: 'commonjs2',
+		path: path.join(__dirname, '../dist/electron')
+	},
+	resolve: {
+		alias: {
+			'@': path.join(__dirname, '../src/renderer'),
+			'vue$': 'vue/dist/vue.esm.js'
+		},
 		extensions: ['.js', '.vue', '.json', '.css', '.node'],
 		modules: [
 			path.resolve(__dirname, '../node_modules'),
 			path.resolve(__dirname, '../src/renderer'),
 		],
-  },
-  target: 'electron-renderer'
+	},
+	target: 'electron-renderer'
 }
 
 /**
  * Adjust rendererConfig for development settings
  */
 if (process.env.NODE_ENV !== 'production') {
-  rendererConfig.plugins.push(
-    new webpack.DefinePlugin({
-      '__static': `"${path.join(__dirname, '../static').replace(/\\/g, '\\\\')}"`
-    })
-  )
+	rendererConfig.plugins.push(
+		new webpack.DefinePlugin({
+			'__static': `"${path.join(__dirname, '../static').replace(/\\/g, '\\\\')}"`
+		})
+	)
 }
 
 /**
  * Adjust rendererConfig for production settings
 */
 if (process.env.NODE_ENV === 'production') {
-  rendererConfig.devtool = ''
+	rendererConfig.devtool = ''
 
-  rendererConfig.plugins.push(
-    new BabelMinifyPlugin(),
-    new CopyWebpackPlugin([
-      {
-        from: path.join(__dirname, '../static'),
-        to: path.join(__dirname, '../dist/electron/static'),
-        ignore: ['.*']
-      }
-    ]),
-    new webpack.DefinePlugin({
-      'process.env.NODE_ENV': '"production"'
-    }),
-    new webpack.LoaderOptionsPlugin({
-      minimize: true
-    })
-  )
+	rendererConfig.plugins.push(
+		new BabelMinifyPlugin(),
+		new CopyWebpackPlugin([
+			{
+				from: path.join(__dirname, '../static'),
+				to: path.join(__dirname, '../dist/electron/static'),
+				ignore: ['.*']
+			}
+		]),
+		new webpack.DefinePlugin({
+			'process.env.NODE_ENV': '"production"'
+		}),
+		new webpack.LoaderOptionsPlugin({
+			minimize: true
+		})
+	)
 }
 
 module.exports = rendererConfig
