+++ src/renderer/providers/assistant/microphone.js
