+++ src/renderer/providers/assistant/player.js
