+++ src/renderer/providers/authentication/google-electron-auth.js
