+++ src/renderer/providers/commands/assistant.js
