+++ src/renderer/providers/commands/audio-control.js
