+++ src/renderer/providers/store/index.js
@@ -1,0 +1,1 @@
+export { default } from './store';
