--- test/unit/specs/LandingPage.spec.js
@@ -1,13 +1,0 @@
-import Vue from 'vue';
-import LandingPage from '@/components/LandingPage';
-
-describe('LandingPage.vue', () => {
-	it('should render correct contents', () => {
-		const vm = new Vue({
-			el: document.createElement('div'),
-			render: h => h(LandingPage),
-		}).$mount();
-
-		expect(vm.$el.querySelector('.title').textContent).to.contain('Welcome to your new project!');
-	});
-});
